[Refactor] : jwt 파일 위치 수정 (#21)
