+++ src/main/java/com/potato/balbambalbam/config/SecurityConfig.java
@@ -1,8 +1,8 @@
 package com.potato.balbambalbam.config;
 
 import com.potato.balbambalbam.data.repository.RefreshRepository;
-import com.potato.balbambalbam.jwt.JWTFilter;
-import com.potato.balbambalbam.jwt.JWTUtil;
+import com.potato.balbambalbam.user.join.jwt.JWTFilter;
+import com.potato.balbambalbam.user.join.jwt.JWTUtil;
 import com.potato.balbambalbam.user.join.jwt.CustomLogoutFilter;
 import com.potato.balbambalbam.user.join.jwt.LoginFilter;
 import org.springframework.context.annotation.Bean;
