+++ src/main/java/com/potato/balbambalbam/user/join/jwt/CustomLogoutFilter.java
@@ -1,7 +1,6 @@
 package com.potato.balbambalbam.user.join.jwt;
 
 import com.potato.balbambalbam.data.repository.RefreshRepository;
-import com.potato.balbambalbam.jwt.JWTUtil;
 import io.jsonwebtoken.ExpiredJwtException;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
@@ -10,8 +9,6 @@ import jakarta.servlet.ServletResponse;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.filter.GenericFilterBean;
 
 import java.io.IOException;
@@ -54,8 +51,10 @@ public class CustomLogoutFilter extends GenericFilterBean {
         }
 
         if (refresh == null) {
+            response.setContentType("text/plain; charset=UTF-8");
             PrintWriter writer = response.getWriter();
             writer.print("refresh 토큰이 없습니다.");
+
             response.setStatus(HttpServletResponse.SC_NOT_FOUND); //404
             return;
         }
@@ -63,6 +62,7 @@ public class CustomLogoutFilter extends GenericFilterBean {
         try {
             jwtUtil.isExpired(refresh);
         } catch (ExpiredJwtException e) {
+            response.setContentType("text/plain; charset=UTF-8");
             PrintWriter writer = response.getWriter();
             writer.print("refresh 토큰이 만료되었습니다.");
 
@@ -72,6 +72,7 @@ public class CustomLogoutFilter extends GenericFilterBean {
 
         String category = jwtUtil.getCategory(refresh);
         if (!category.equals("refresh")) {
+            response.setContentType("text/plain; charset=UTF-8");
             PrintWriter writer = response.getWriter();
             writer.print("refresh 토큰이 아닙니다.");
 
@@ -81,10 +82,11 @@ public class CustomLogoutFilter extends GenericFilterBean {
 
         Boolean isExist = refreshRepository.existsByRefresh(refresh);
         if (!isExist) {
+            response.setContentType("text/plain; charset=UTF-8");
             PrintWriter writer = response.getWriter();
             writer.print("refresh 토큰이 데이터베이스에 없습니다.");
 
-            response.setStatus(HttpServletResponse.SC_BAD_REQUEST); //400
+            response.setStatus(HttpServletResponse.SC_REQUEST_TIMEOUT); //400
             return;
         }
