+++ src/main/java/com/potato/balbambalbam/user/join/jwt/JWTFilter.java
@@ -1,7 +1,8 @@
-package com.potato.balbambalbam.jwt;
+package com.potato.balbambalbam.user.join.jwt;
 
 import com.potato.balbambalbam.data.entity.User;
 import com.potato.balbambalbam.user.join.dto.CustomUserDetails;
+import com.potato.balbambalbam.user.join.jwt.JWTUtil;
 import io.jsonwebtoken.ExpiredJwtException;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
